safety: L4 word-aligned diag readback, bram_safety strip, loopback env, L4 scenarios (#71)
SafetyDiag state fields (health/degraded_class/fault/can_state) word-aligned
(32-bit) so runner readback rules parse them as whole words. flash.py strips
and clears .bram_safety (was loadable -> upload verify fail; stale counters
broke eq rules). Add env:firmware-loopback (-DCAN_LOOPBACK) for the L2 leg.
Add L4 scenarios safety-acquire-bus/loopback/marker + owner checklist.

Verified: host 84/84; bench L4 legs PASS (bus: Degraded->Fault(DegradedTimeout)
60s + force-stop; loopback: zero-frame stop emission; marker: boot1+boot2
Fault(CrashMarker), power-cycle != ack).

+++ adapters/can_bus.cpp
@@ -83,7 +83,7 @@ void CanBus::init(bool loopback, SafetyDiag* diag)
     (void)can_init(loopback);
     if (m_diag != nullptr)
     {
-        m_diag->can_state = error_state();
+        m_diag->can_state = static_cast<std::uint32_t>(error_state());
     }
 }
 
@@ -239,7 +239,7 @@ void CanBus::recover_bus_off()
     if (m_diag != nullptr)
     {
         ++m_diag->can_bus_off_recoveries;
-        m_diag->can_state = error_state();
+        m_diag->can_state = static_cast<std::uint32_t>(error_state());
     }
 }
